глобальный сейв

+++ client/src/modules/centerFrame/history/components/HistoryList.tsx
@@ -11,14 +11,22 @@ import { useGetImages } from '../../tabletours/hooks/getImages'
 export const HistoryList: React.FC<{ data: Tournament[] }> = ({ data }) => {
     const { getImage } = useGetImages()
     const [rows, setRows] = useState<JSX.Element[]>([]);
+    const [admin, set_admin] = useState<boolean>(false)
     const dispatch = useDispatch()
     const catchsList = useSelector((state: RootState) => state.slice.catchsList);
     const idClickTour = useSelector((state: RootState) => state.slice.idClickTour)
     const celevoys = data.find(e => e.id === idClickTour)
+    const userStatus = useSelector((state: RootState) => state.slice.userStatus);
+    const userID = userStatus?.user?.id
+    const created = userStatus.tournament.find(e => e.id === idClickTour)
     console.log(catchsList)
 
 
+
+
     const handler = (catchItem: ExtendedBigFish) => {
+        console.log(userID, catchItem.idUser, created?.created_by)
+        if (userID !== catchItem.idUser && userID !== created?.created_by) return
         dispatch(set_stateModalWindow({ type: 'deleteForm', status: true }))
         dispatch(set_deleteIdCatch(catchItem))
     }
@@ -50,35 +58,47 @@ export const HistoryList: React.FC<{ data: Tournament[] }> = ({ data }) => {
         }))
     }
     useEffect(() => {
-        console.log('тут ведь')
+        console.log('тут ведь');
         const fetchRows = async () => {
             const newRows = await Promise.all(
-                catchsList.map(async e => {
+                catchsList.map(async (e) => {
                     const time = formatUnixTime(Number(e.data));
-                    let imageUrl = e.urlFoto ? await getImage(e.urlFoto) : 'Без фото';
+                    let imageUrl = 'Без фото';
+
+                    if (e.urlFoto) {
+                        try {
+                            // Получение изображения с кэшированием внутри хука
+                            imageUrl = await getImage(e.urlFoto);
+                        } catch (error) {
+                            console.error(`Ошибка загрузки изображения ${e.urlFoto}`, error);
+                        }
+                    }
+
+                    const admin = userID === e.idUser || created?.created_by === userID
+
                     return (
                         <tr className='rows_list_catch' key={e.data}>
-                            <td>{e.name_user}</td>
-                            <td>{e.name_reservour}</td>
-                            <td>{e.name_fish}</td>
-                            <td>{e.name_day}</td>
-                            <td>{e.name_type}</td>
-                            <td>{e.name_bait}</td>
-                            <td>{e.weight}</td>
-                            <td className='edit' onClick={() => editHandler(e)}>{time}</td>
-                            <td className='icon_fish_foto' style={{
-                                backgroundImage: `url(${imageUrl})`,
-                            }}></td>
-                            <td className='delete_catch' onClick={() => handler(e)}><FaTimes className='icon_del' /></td>
+                            <td className="row_history_catch">{e.name_user}</td>
+                            <td className="row_history_catch">{e.name_fish}</td>
+                            <td className="row_history_catch">{e.name_type}</td>
+                            <td className="row_history_catch">{e.weight}</td>
+                            <td className='row_history_catch edit' onClick={() => editHandler(e)}>{time}</td>
+                            <td
+                                className='row_history_catch icon_fish_foto'
+                                style={{
+                                    backgroundImage: imageUrl !== 'Без фото' ? `url(${imageUrl})` : undefined,
+                                }}
+                            ></td>
+                            <td className='row_history_catch delete_catch' onClick={() => handler(e)}>{admin && <FaTimes className='icon_del' />}</td>
                         </tr>
                     );
                 })
             );
-            setRows(newRows); // Обновляем состояние с полученными строками
+            setRows(newRows);
         };
 
-        fetchRows(); // Вызовем асинхронную функцию
-    }, [catchsList]); // Зависимость по catchsList
+        fetchRows();
+    }, [catchsList]);
 
     console.log(rows)
     return <div className="card_list_history">
@@ -88,16 +108,13 @@ export const HistoryList: React.FC<{ data: Tournament[] }> = ({ data }) => {
         <div className="container_table_history">
             {rows.length !== 0 ? <table className='styled_table'>
                 <thead><tr className="history_header">
-                    <th>Участник</th>
-                    <th>Водоём</th>
-                    <th>Вид рыбы</th>
-                    <th>Время суток</th>
-                    <th>Тип ловли</th>
-                    <th>Приманка</th>
-                    <th>Вес</th>
-                    <th>Время</th>
-                    <th>Фото</th>
-                    <th>Удалить</th>
+                    <th className="header_history_catch">Участник</th>
+                    <th className="header_history_catch">Вид рыбы</th>
+                    <th className="header_history_catch">Тип ловли</th>
+                    <th className="header_history_catch">Вес</th>
+                    <th className="header_history_catch">Время</th>
+                    <th className="header_history_catch">Фото</th>
+                    <th className="header_history_catch">Удалить</th>
                 </tr>
                 </thead>
                 <tbody>{rows}</tbody>
